requirements.txt
matplotlib
pytesseract
pillow
numpy
